refactor: CLEARDS01 to DSPERF02
