+++ rpgJavaInterpreter-core/src/test/resources/DSPERF02.rpgle
